refactor(web): drop obsolete functions page
Decompose common components.
