+++ web/src/pages/builtin/_shared/lane-editor/FlowLink.tsx
@@ -4,5 +4,5 @@ import React from 'react';
  * Short dashed connector arrow between module cards in a lane track.
  */
 export const FlowLink = (): React.JSX.Element => (
-    <div className="fng-flow-link" aria-hidden="true" />
+    <div className="fn-flow-link" aria-hidden="true" />
 );
