+++ web/src/pages/builtin/_shared/lane-editor/Sparkline.tsx
