--- web/src/pages/builtin/functions-ng/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FunctionsNgPage';
