+++ web/src/pages/builtin/functions/dialogs/index.ts
@@ -1,11 +1,2 @@
 export { CreateFunctionDialog } from './CreateFunctionDialog';
 export type { CreateFunctionDialogProps } from './CreateFunctionDialog';
-
-export { WeightEditorDialog } from './WeightEditorDialog';
-export type { WeightEditorDialogProps } from './WeightEditorDialog';
-
-export { ModuleEditorDialog } from './ModuleEditorDialog';
-export type { ModuleEditorDialogProps } from './ModuleEditorDialog';
-
-export { SingleWeightEditorDialog } from './SingleWeightEditorDialog';
-export type { SingleWeightEditorDialogProps } from './SingleWeightEditorDialog';
