+++ web/src/pages/builtin/functions/index.ts
@@ -1,12 +1,1 @@
-export * from './types';
-export * from './utils';
-export * from './hooks';
-export { FunctionGraph } from './FunctionGraph';
-export { FunctionCard } from './FunctionCard';
-export { FunctionPageHeader } from './FunctionPageHeader';
-export * from './dialogs';
-export {
-    InputNode as InputNodeComponent,
-    OutputNode as OutputNodeComponent,
-    ModuleNode as ModuleNodeComponent,
-} from './nodes';
+export { default } from './FunctionsPage';
