--- web/src/pages/builtin/functions/nodes/index.ts
@@ -1,4 +1,0 @@
-export { InputNode } from './InputNode';
-export { OutputNode } from './OutputNode';
-export { ModuleNode } from './ModuleNode';
-
