+++ web/src/pages/builtin/functions/reducer.ts
