+++ web/src/pages/builtin/pipelines/nodes/OutputNode.tsx
