+++ web/src/pages/builtin/pipelines/nodes/index.ts
@@ -1,5 +1,4 @@
 export { FunctionRefNode } from './FunctionRefNode';
 export { CounterEdge } from './CounterEdge';
-
-// Re-export Input/Output nodes from functions
-export { InputNode, OutputNode } from '../../functions/nodes';
+export { InputNode } from './InputNode';
+export { OutputNode } from './OutputNode';
