+++ web/src/types.ts
@@ -3,7 +3,6 @@ import { createContext, useContext } from 'react';
 export const PAGE_IDS = [
     'builtin/inspect',
     'builtin/functions',
-    'builtin/functions-ng',
     'builtin/pipelines',
     'builtin/devices',
     'modules/forward',
